Move errors-specific copy to its own file
+ Capitalize ID heading
